make physics generalized, create more characters
physics has been moved to a script so that it can be added to new
characters with just one line.

added new characters with placeholder sprites and some unique stats.

also, you can shorthop now. tapping the jump button and releasing it
within five frames is treated as a shorthop rather than a full jump.